refactor(kilo-pass): isolate affiliate sale enqueue
